Move test-asyncio-simplerouter under tests/ dir
Also, make some test components into fixtures.
